+++ tests/unit/sagemaker/workflow/helpers.py
@@ -10,11 +10,10 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-# language governing permissions and limitations under the License.
 """Helper methods for testing."""
 from __future__ import absolute_import
 
-from sagemaker.workflow import Properties
+from sagemaker.workflow.properties import Properties
 from sagemaker.workflow.steps import Step, StepTypeEnum