+++ tests/unit/sagemaker/workflow/test_training_step.py
@@ -10,16 +10,20 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-# language governing permissions and limitations under the License.
 from __future__ import absolute_import
 
 import os
 import json
+from mock import Mock, PropertyMock
+import re
 
 import pytest
-import sagemaker
 import warnings
 
+from sagemaker import Processor, Model
+from sagemaker.parameter import IntegerParameter
+from sagemaker.transformer import Transformer
+from sagemaker.tuner import HyperparameterTuner
 from sagemaker.workflow.pipeline_context import PipelineSession
 from sagemaker.workflow.parameters import ParameterString
 
@@ -48,12 +52,14 @@ from sagemaker.amazon.randomcutforest import RandomCutForest
 from sagemaker.amazon.ntm import NTM
 from sagemaker.amazon.object2vec import Object2Vec
 
-from tests.integ import DATA_DIR
+from tests.unit import DATA_DIR
 
 from sagemaker.inputs import TrainingInput
 from tests.unit.sagemaker.workflow.helpers import CustomStep
 
 REGION = "us-west-2"
+BUCKET = "my-bucket"
+ROLE = "DummyRole"
 IMAGE_URI = "fakeimage"
 MODEL_NAME = "gisele"
 DUMMY_LOCAL_SCRIPT_PATH = os.path.join(DATA_DIR, "dummy_script.py")
@@ -63,18 +69,48 @@ INSTANCE_TYPE = "ml.m4.xlarge"
 
 
 @pytest.fixture
-def pipeline_session():
-    return PipelineSession()
+def client():
+    """Mock client.
+
+    Considerations when appropriate:
+
+        * utilize botocore.stub.Stubber
+        * separate runtime client from client
+    """
+    client_mock = Mock()
+    client_mock._client_config.user_agent = (
+        "Boto3/1.14.24 Python/3.8.5 Linux/5.4.0-42-generic Botocore/1.17.24 Resource"
+    )
+    return client_mock
 
 
 @pytest.fixture
-def bucket(pipeline_session):
-    return pipeline_session.default_bucket()
+def boto_session(client):
+    role_mock = Mock()
+    type(role_mock).arn = PropertyMock(return_value=ROLE)
+
+    resource_mock = Mock()
+    resource_mock.Role.return_value = role_mock
+
+    session_mock = Mock(region_name=REGION)
+    session_mock.resource.return_value = resource_mock
+    session_mock.client.return_value = client
+
+    return session_mock
 
 
 @pytest.fixture
-def training_input(bucket):
-    return TrainingInput(s3_data=f"s3://{bucket}/my-training-input")
+def pipeline_session(boto_session, client):
+    return PipelineSession(
+        boto_session=boto_session,
+        sagemaker_client=client,
+        default_bucket=BUCKET,
+    )
+
+
+@pytest.fixture
+def training_input():
+    return TrainingInput(s3_data=f"s3://{BUCKET}/my-training-input")
 
 
 @pytest.fixture
@@ -86,7 +122,7 @@ def test_training_step_with_estimator(pipeline_session, training_input, hyperpar
     custom_step1 = CustomStep("TestStep")
     custom_step2 = CustomStep("SecondTestStep")
     estimator = Estimator(
-        role=sagemaker.get_execution_role(),
+        role=ROLE,
         instance_count=1,
         instance_type=INSTANCE_TYPE,
         sagemaker_session=pipeline_session,
@@ -121,17 +157,18 @@ def test_training_step_with_estimator(pipeline_session, training_input, hyperpar
         "DisplayName": "MyTrainingStep",
         "Type": "Training",
         "DependsOn": ["TestStep", "SecondTestStep"],
-        "Arguments": step_args,
+        "Arguments": step_args.args,
     }
     assert step.properties.TrainingJobName.expr == {"Get": "Steps.MyTrainingStep.TrainingJobName"}
 
 
 def test_estimator_with_parameterized_output(pipeline_session, training_input):
     output_path = ParameterString(name="OutputPath")
+    # XGBoost
     estimator = XGBoost(
         framework_version="1.3-1",
         py_version="py3",
-        role=sagemaker.get_execution_role(),
+        role=ROLE,
         instance_type=INSTANCE_TYPE,
         instance_count=1,
         entry_point=DUMMY_LOCAL_SCRIPT_PATH,
@@ -139,21 +176,48 @@ def test_estimator_with_parameterized_output(pipeline_session, training_input):
         sagemaker_session=pipeline_session,
     )
     step_args = estimator.fit(inputs=training_input)
-    step = TrainingStep(
-        name="MyTrainingStep",
+    step1 = TrainingStep(
+        name="MyTrainingStep1",
+        step_args=step_args,
+        description="TrainingStep description",
+        display_name="MyTrainingStep",
+    )
+
+    # TensorFlow
+    # If model_dir is None and output_path is a pipeline variable
+    # a default model_dir will be generated with default bucket
+    estimator = TensorFlow(
+        framework_version="2.4.1",
+        py_version="py37",
+        role=ROLE,
+        instance_type=INSTANCE_TYPE,
+        instance_count=1,
+        entry_point=DUMMY_LOCAL_SCRIPT_PATH,
+        output_path=output_path,
+        sagemaker_session=pipeline_session,
+    )
+    step_args = estimator.fit(inputs=training_input)
+    step2 = TrainingStep(
+        name="MyTrainingStep2",
         step_args=step_args,
         description="TrainingStep description",
         display_name="MyTrainingStep",
     )
     pipeline = Pipeline(
         name="MyPipeline",
-        steps=[step],
+        steps=[step1, step2],
+        parameters=[output_path],
         sagemaker_session=pipeline_session,
     )
-    step_def = json.loads(pipeline.definition())["Steps"][0]
-    assert step_def["Arguments"]["OutputDataConfig"]["S3OutputPath"] == {
-        "Get": "Parameters.OutputPath"
-    }
+    step_defs = json.loads(pipeline.definition())["Steps"]
+    for step_def in step_defs:
+        assert step_def["Arguments"]["OutputDataConfig"]["S3OutputPath"] == {
+            "Get": "Parameters.OutputPath"
+        }
+        if step_def["Name"] != "MyTrainingStep2":
+            continue
+        model_dir = step_def["Arguments"]["HyperParameters"]["model_dir"]
+        assert re.match(rf'"s3://{BUCKET}/.*/model"', model_dir)
 
 
 @pytest.mark.parametrize(
@@ -164,11 +228,11 @@ def test_estimator_with_parameterized_output(pipeline_session, training_input):
             py_version="py3",
             instance_type=INSTANCE_TYPE,
             instance_count=1,
-            role=sagemaker.get_execution_role(),
+            role=ROLE,
             entry_point=DUMMY_LOCAL_SCRIPT_PATH,
         ),
         PyTorch(
-            role=sagemaker.get_execution_role(),
+            role=ROLE,
             instance_type=INSTANCE_TYPE,
             instance_count=1,
             framework_version="1.8.0",
@@ -176,7 +240,7 @@ def test_estimator_with_parameterized_output(pipeline_session, training_input):
             entry_point=DUMMY_LOCAL_SCRIPT_PATH,
         ),
         TensorFlow(
-            role=sagemaker.get_execution_role(),
+            role=ROLE,
             instance_type=INSTANCE_TYPE,
             instance_count=1,
             framework_version="2.0",
@@ -186,7 +250,7 @@ def test_estimator_with_parameterized_output(pipeline_session, training_input):
         HuggingFace(
             transformers_version="4.6",
             pytorch_version="1.7",
-            role=sagemaker.get_execution_role(),
+            role=ROLE,
             instance_type="ml.p3.2xlarge",
             instance_count=1,
             py_version="py36",
@@ -195,7 +259,7 @@ def test_estimator_with_parameterized_output(pipeline_session, training_input):
         XGBoost(
             framework_version="1.3-1",
             py_version="py3",
-            role=sagemaker.get_execution_role(),
+            role=ROLE,
             instance_type=INSTANCE_TYPE,
             instance_count=1,
             entry_point=DUMMY_LOCAL_SCRIPT_PATH,
@@ -203,7 +267,7 @@ def test_estimator_with_parameterized_output(pipeline_session, training_input):
         MXNet(
             framework_version="1.4.1",
             py_version="py3",
-            role=sagemaker.get_execution_role(),
+            role=ROLE,
             instance_type=INSTANCE_TYPE,
             instance_count=1,
             entry_point=DUMMY_LOCAL_SCRIPT_PATH,
@@ -213,12 +277,12 @@ def test_estimator_with_parameterized_output(pipeline_session, training_input):
             toolkit=RLToolkit.RAY,
             framework=RLFramework.TENSORFLOW,
             toolkit_version="0.8.5",
-            role=sagemaker.get_execution_role(),
+            role=ROLE,
             instance_type=INSTANCE_TYPE,
             instance_count=1,
         ),
         Chainer(
-            role=sagemaker.get_execution_role(),
+            role=ROLE,
             entry_point=DUMMY_LOCAL_SCRIPT_PATH,
             use_mpi=True,
             num_processes=4,
@@ -254,7 +318,7 @@ def test_training_step_with_framework_estimator(
     assert json.loads(pipeline.definition())["Steps"][0] == {
         "Name": "MyTrainingStep",
         "Type": "Training",
-        "Arguments": step_args,
+        "Arguments": step_args.args,
     }
 
 
@@ -275,13 +339,13 @@ def test_training_step_with_framework_estimator(
 )
 def test_training_step_with_algorithm_base(algo_estimator, pipeline_session):
     estimator = algo_estimator(
-        role=sagemaker.get_execution_role(),
+        role=ROLE,
         instance_type=INSTANCE_TYPE,
         instance_count=1,
         sagemaker_session=pipeline_session,
     )
     data = RecordSet(
-        "s3://{}/{}".format(pipeline_session.default_bucket(), "dummy"),
+        "s3://{}/{}".format(BUCKET, "dummy"),
         num_records=1000,
         feature_dim=128,
         channel="train",
@@ -311,6 +375,71 @@ def test_training_step_with_algorithm_base(algo_estimator, pipeline_session):
     assert json.loads(pipeline.definition())["Steps"][0] == {
         "Name": "MyTrainingStep",
         "Type": "Training",
-        "Arguments": step_args,
+        "Arguments": step_args.args,
     }
     assert step.properties.TrainingJobName.expr == {"Get": "Steps.MyTrainingStep.TrainingJobName"}
+
+
+@pytest.mark.parametrize(
+    "inputs",
+    [
+        (
+            Processor(
+                image_uri=IMAGE_URI,
+                role=ROLE,
+                instance_count=1,
+                instance_type=INSTANCE_TYPE,
+            ),
+            dict(target_fun="run", func_args={}),
+        ),
+        (
+            Transformer(
+                model_name="model_name",
+                instance_type="ml.m5.xlarge",
+                instance_count=1,
+                output_path="s3://Transform",
+            ),
+            dict(
+                target_fun="transform",
+                func_args=dict(data="s3://data", job_name="test"),
+            ),
+        ),
+        (
+            HyperparameterTuner(
+                estimator=Estimator(
+                    role=ROLE,
+                    instance_count=1,
+                    instance_type=INSTANCE_TYPE,
+                    image_uri=IMAGE_URI,
+                ),
+                objective_metric_name="test:acc",
+                hyperparameter_ranges={"batch-size": IntegerParameter(64, 128)},
+            ),
+            dict(target_fun="fit", func_args={}),
+        ),
+        (
+            Model(
+                image_uri=IMAGE_URI,
+                role=ROLE,
+            ),
+            dict(target_fun="create", func_args={}),
+        ),
+    ],
+)
+def test_insert_wrong_step_args_into_training_step(inputs, pipeline_session):
+    downstream_obj, target_func_cfg = inputs
+    if isinstance(downstream_obj, HyperparameterTuner):
+        downstream_obj.estimator.sagemaker_session = pipeline_session
+    else:
+        downstream_obj.sagemaker_session = pipeline_session
+    func_name = target_func_cfg["target_fun"]
+    func_args = target_func_cfg["func_args"]
+    step_args = getattr(downstream_obj, func_name)(**func_args)
+
+    with pytest.raises(ValueError) as error:
+        TrainingStep(
+            name="MyTrainingStep",
+            step_args=step_args,
+        )
+
+    assert "The step_args of TrainingStep must be obtained from estimator.fit()" in str(error.value)