+++ tests/unit/sagemaker/workflow/test_tuning_step.py
@@ -10,15 +10,17 @@
 # distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
-# language governing permissions and limitations under the License.
 from __future__ import absolute_import
 
 import os
 import json
+from mock import Mock, PropertyMock
 import pytest
 import warnings
 
-import sagemaker
+from sagemaker import Model, Processor
+from sagemaker.estimator import Estimator
+from sagemaker.transformer import Transformer
 from sagemaker.workflow.pipeline_context import PipelineSession
 
 from sagemaker.workflow.steps import TuningStep
@@ -31,9 +33,51 @@ from sagemaker.pytorch.estimator import PyTorch
 from tests.unit import DATA_DIR
 
 
+REGION = "us-west-2"
+BUCKET = "my-bucket"
+ROLE = "DummyRole"
+IMAGE_URI = "fakeimage"
+INSTANCE_TYPE = "ml.m4.xlarge"
+
+
 @pytest.fixture
-def pipeline_session():
-    return PipelineSession()
+def client():
+    """Mock client.
+
+    Considerations when appropriate:
+
+        * utilize botocore.stub.Stubber
+        * separate runtime client from client
+    """
+    client_mock = Mock()
+    client_mock._client_config.user_agent = (
+        "Boto3/1.14.24 Python/3.8.5 Linux/5.4.0-42-generic Botocore/1.17.24 Resource"
+    )
+    return client_mock
+
+
+@pytest.fixture
+def boto_session(client):
+    role_mock = Mock()
+    type(role_mock).arn = PropertyMock(return_value=ROLE)
+
+    resource_mock = Mock()
+    resource_mock.Role.return_value = role_mock
+
+    session_mock = Mock(region_name=REGION)
+    session_mock.resource.return_value = resource_mock
+    session_mock.client.return_value = client
+
+    return session_mock
+
+
+@pytest.fixture
+def pipeline_session(boto_session, client):
+    return PipelineSession(
+        boto_session=boto_session,
+        sagemaker_client=client,
+        default_bucket=BUCKET,
+    )
 
 
 @pytest.fixture
@@ -47,7 +91,7 @@ def test_tuning_step_with_single_algo_tuner(pipeline_session, entry_point):
 
     pytorch_estimator = PyTorch(
         entry_point=entry_point,
-        role=sagemaker.get_execution_role(),
+        role=ROLE,
         framework_version="1.5.0",
         py_version="py3",
         instance_count=1,
@@ -93,14 +137,14 @@ def test_tuning_step_with_single_algo_tuner(pipeline_session, entry_point):
     assert json.loads(pipeline.definition())["Steps"][0] == {
         "Name": "MyTuningStep",
         "Type": "Tuning",
-        "Arguments": step_args,
+        "Arguments": step_args.args,
     }
 
 
 def test_tuning_step_with_multi_algo_tuner(pipeline_session, entry_point):
     pytorch_estimator = PyTorch(
         entry_point=entry_point,
-        role=sagemaker.get_execution_role(),
+        role=ROLE,
         framework_version="1.5.0",
         py_version="py3",
         instance_count=1,
@@ -156,5 +200,66 @@ def test_tuning_step_with_multi_algo_tuner(pipeline_session, entry_point):
     assert json.loads(pipeline.definition())["Steps"][0] == {
         "Name": "MyTuningStep",
         "Type": "Tuning",
-        "Arguments": step_args,
+        "Arguments": step_args.args,
     }
+
+
+@pytest.mark.parametrize(
+    "inputs",
+    [
+        (
+            Transformer(
+                model_name="model_name",
+                instance_type="ml.m5.xlarge",
+                instance_count=1,
+                output_path="s3://Transform",
+            ),
+            dict(
+                target_fun="transform",
+                func_args=dict(data="s3://data", job_name="test"),
+            ),
+        ),
+        (
+            Estimator(
+                role=ROLE,
+                instance_count=1,
+                instance_type=INSTANCE_TYPE,
+                image_uri=IMAGE_URI,
+            ),
+            dict(
+                target_fun="fit",
+                func_args={},
+            ),
+        ),
+        (
+            Processor(
+                image_uri=IMAGE_URI,
+                role=ROLE,
+                instance_count=1,
+                instance_type=INSTANCE_TYPE,
+            ),
+            dict(target_fun="run", func_args={}),
+        ),
+        (
+            Model(
+                image_uri=IMAGE_URI,
+                role=ROLE,
+            ),
+            dict(target_fun="create", func_args={}),
+        ),
+    ],
+)
+def test_insert_wrong_step_args_into_tuning_step(inputs, pipeline_session):
+    downstream_obj, target_func_cfg = inputs
+    downstream_obj.sagemaker_session = pipeline_session
+    func_name = target_func_cfg["target_fun"]
+    func_args = target_func_cfg["func_args"]
+    step_args = getattr(downstream_obj, func_name)(**func_args)
+
+    with pytest.raises(ValueError) as error:
+        TuningStep(
+            name="MyTuningStep",
+            step_args=step_args,
+        )
+
+    assert "The step_args of TuningStep must be obtained from tuner.fit()" in str(error.value)